+++ NEWS.md
@@ -1,3 +1,9 @@
+# peskas.timor.portal 0.1.0
+
+## New features
+
+- Added a leaflet dynamic map displaying fishery indicatory along Timor coast. 
+
 # peskas.timor.portal 0.0.0.9000
 
 * Added a `NEWS.md` file to track changes to the package.